rename node_settings file
