+++ silkworm/node/backend/ethereum_backend.hpp
@@ -24,7 +24,7 @@
 #include <silkworm/core/chain/config.hpp>
 #include <silkworm/core/common/base.hpp>
 #include <silkworm/node/backend/state_change_collection.hpp>
-#include <silkworm/node/common/settings.hpp>
+#include <silkworm/node/common/node_settings.hpp>
 #include <silkworm/sentry/api/common/sentry_client.hpp>
 
 namespace silkworm {
