+++ silkworm/node/node.hpp
@@ -20,7 +20,7 @@
 
 #include <silkworm/infra/concurrency/task.hpp>
 
-#include <silkworm/node/common/settings.hpp>
+#include <silkworm/node/common/node_settings.hpp>
 #include <silkworm/node/db/mdbx.hpp>
 #include <silkworm/node/settings.hpp>
 #include <silkworm/node/stagedsync/local_client.hpp>
