+++ silkworm/sync/sync_pos.hpp
@@ -20,7 +20,7 @@
 
 #include <silkworm/infra/common/log.hpp>
 #include <silkworm/infra/concurrency/active_component.hpp>
-#include <silkworm/node/common/settings.hpp>
+#include <silkworm/node/common/node_settings.hpp>
 #include <silkworm/node/stagedsync/client.hpp>
 #include <silkworm/rpc/types/execution_payload.hpp>
 #include <silkworm/sync/internals/chain_fork_view.hpp>
